Moves .css files to correct folder and updates paths
